requirements.txt
absl-py==0.10.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
easyocr==1.1.10
editdistance==0.5.3
efficientnet==1.0.0
essential-generators==0.9.2
fonttools==4.16.1
future==0.18.2
gast==0.3.3
google-api-core==1.22.4
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-cloud-vision==2.0.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
imutils==0.5.3
Keras-Applications==1.0.8
keras-ocr==0.8.5
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
libcst==0.3.13
Markdown==3.3.2
matplotlib==3.3.2
mypy-extensions==0.4.3
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
pandas==1.1.3
Pillow==8.0.0
proto-plus==1.11.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.2.0
pyparsing==2.4.7
pytesseract==0.3.6
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.3
Shapely==1.7.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tesseract==0.1.3
tesseract-ocr==0.0.1
tifffile==2020.10.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
validators==0.18.1
Werkzeug==1.0.1
wrapt==1.12.1
